Refactor project structure - move RealmClientHandler to client_handler module
