+++ warden_daemon/src/client_handler.rs
@@ -1,1 +1,2 @@
 pub mod client_command_handler;
+pub mod realm_client_handler;
