+++ warden_daemon/src/fabric/application_fabric.rs
@@ -4,8 +4,7 @@ use tokio::sync::Mutex;
 
 use crate::managers::{
     application::{Application, ApplicationConfig, ApplicationCreator},
-    application_manager::ApplicationManager,
-    realm_manager::RealmClient,
+    application_manager::ApplicationManager, realm_client::RealmClient,
 };
 
 pub struct ApplicationFabric {}
@@ -20,8 +19,8 @@ impl ApplicationCreator for ApplicationFabric {
     fn create_application(
         &self,
         config: ApplicationConfig,
-        realm_client: Arc<Mutex<Box<dyn RealmClient + Send + Sync>>>,
+        realm_client_handler: Arc<Mutex<Box<dyn RealmClient + Send + Sync>>>,
     ) -> Box<dyn Application + Send + Sync> {
-        Box::new(ApplicationManager::new(config, realm_client))
+        Box::new(ApplicationManager::new(config, realm_client_handler))
     }
 }
