+++ warden_daemon/src/managers.rs
@@ -1,8 +1,8 @@
 pub mod application;
 pub mod application_manager;
 pub mod realm;
-pub mod realm_client;
 pub mod realm_configuration;
 pub mod realm_manager;
 pub mod warden;
 pub mod warden_manager;
+pub mod realm_client;
