+++ warden_daemon/src/managers/application_manager.rs
@@ -3,24 +3,23 @@ use std::sync::Arc;
 use tokio::sync::Mutex;
 
 use super::{
-    application::{Application, ApplicationConfig, ApplicationError},
-    realm_manager::RealmClient,
+    application::{Application, ApplicationConfig, ApplicationError}, realm_client::RealmClient
 };
 
 pub struct ApplicationManager {
     config: ApplicationConfig,
-    realm_client: Arc<Mutex<Box<dyn RealmClient + Send + Sync>>>,
+    realm_client_handler: Arc<Mutex<Box<dyn RealmClient + Send + Sync>>>,
     new_config: Option<ApplicationConfig>,
 }
 
 impl ApplicationManager {
     pub fn new(
         config: ApplicationConfig,
-        realm_client: Arc<Mutex<Box<dyn RealmClient + Send + Sync>>>,
+        realm_client_handler: Arc<Mutex<Box<dyn RealmClient + Send + Sync>>>,
     ) -> Self {
         ApplicationManager {
             config,
-            realm_client,
+            realm_client_handler,
             new_config: None,
         }
     }
@@ -29,7 +28,7 @@ impl ApplicationManager {
 #[async_trait]
 impl Application for ApplicationManager {
     async fn stop(&mut self) -> Result<(), ApplicationError> {
-        self.realm_client
+        self.realm_client_handler
             .lock()
             .await
             .stop_application(&self.config.uuid)
@@ -38,7 +37,7 @@ impl Application for ApplicationManager {
         Ok(())
     }
     async fn start(&mut self) -> Result<(), ApplicationError> {
-        self.realm_client
+        self.realm_client_handler
             .lock()
             .await
             .start_application(&self.config.uuid)
